refactor(wallet-filter-research): update wallet data handling and taxonomy references
- Revise wallet comparison logic to utilize `walletData` records for identifying new, renamed, or changed wallet products.
- Clarify the process for researching wallets, emphasizing the use of canonical records and the removal of override layers.
- Update documentation to reflect changes in wallet icon handling, company ID mapping, and taxonomy management.
- Ensure consistency between wallet data and taxonomy definitions, with a focus on maintaining synchronization between `wallet-data.ts` and `taxonomy.ts`.
