+++ apps/web/src/app/[locale]/wallets/WalletDirectory.module.scss
