--- apps/web/src/components/wallets/WalletCard.module.scss
@@ -1,86 +1,0 @@
-.wallet {
-  display: flex;
-  flex-flow: column wrap;
-  background-color: #1d1a23;
-  border-radius: 16px;
-  padding: 30px;
-  gap: 15px;
-  transition: all 0.3s ease;
-
-  &:hover {
-    transform: scale(1.05);
-    background-color: #2b2634;
-  }
-}
-
-.wallet-card-header {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-between;
-  align-items: flex-start;
-}
-
-.wallet-card-icon {
-  width: 60px;
-  height: auto;
-}
-
-.wallet-card-badge {
-  display: inline-block;
-  background-color: #2b2634;
-  color: #9945ff;
-  border: 1px solid #3d3650;
-  border-radius: 6px;
-  padding: 3px 8px;
-  font-size: 10px;
-  font-family: "Diatype", monotype;
-  line-height: 14px;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-  white-space: nowrap;
-}
-
-.wallet-card-title {
-  font-size: 28px;
-  line-height: 33px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  margin-bottom: 0;
-}
-
-.wallet-card-body {
-  font-size: 16px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  line-height: 20px;
-  color: #6c6a81;
-  margin-bottom: 0;
-}
-
-.wallet-card-body--alt {
-  font-size: 16px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  line-height: 20px;
-  margin-bottom: 0;
-  color: #fff;
-}
-
-.wallet-card-view-details {
-  font-size: 14px;
-  font-family: "Diatype", monotype;
-  line-height: 20px;
-  letter-spacing: 0.5px;
-  text-transform: uppercase;
-  text-align: left;
-  margin-bottom: 0;
-  color: #fff;
-
-  span {
-    color: #9945ff;
-    font-size: 20px;
-
-    svg {
-      margin-left: 5px;
-      width: 20px;
-      height: 20px;
-    }
-  }
-}
