--- apps/web/src/components/wallets/WalletCard.tsx
@@ -1,84 +1,0 @@
-import { useTranslations } from "next-intl";
-import Image from "next/image";
-import styles from "./WalletCard.module.scss";
-
-const CATEGORY_LABELS = {
-  consumer: "Wallet",
-  infrastructure: "Developer Tool",
-  hardware: "Hardware",
-  institutional: "Institutional",
-  payments: "Payments",
-};
-
-interface WalletCardProps {
-  index: number;
-  walletImage: string;
-  name: string;
-  body: string;
-  websiteUrl: string;
-  category: keyof typeof CATEGORY_LABELS;
-}
-
-const WalletCard = ({
-  index,
-  walletImage,
-  name,
-  body,
-  websiteUrl,
-  category,
-}: WalletCardProps) => {
-  const t = useTranslations();
-
-  return (
-    <article data-index={index} className={styles["wallet"]}>
-      <div className={styles["wallet-card-header"]}>
-        <Image
-          src={walletImage}
-          width={60}
-          height={60}
-          alt={name}
-          className={styles["wallet-card-icon"]}
-          suppressHydrationWarning
-        ></Image>
-        {category !== "consumer" && CATEGORY_LABELS[category] && (
-          <span className={styles["wallet-card-badge"]}>
-            {CATEGORY_LABELS[category]}
-          </span>
-        )}
-      </div>
-      <h3 className={styles["wallet-card-title"]} suppressHydrationWarning>
-        {name}
-      </h3>
-      <p className={styles["wallet-card-body"]} suppressHydrationWarning>
-        {body}
-      </p>
-      <a
-        href={websiteUrl}
-        className={styles["wallet-card-view-details"]}
-        target="_blank"
-        rel="noreferrer"
-        suppressHydrationWarning
-      >
-        {t("wallets.card.view-details")}
-        <span>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            strokeWidth="1.5"
-            stroke="currentColor"
-            className="w-[15px] h-[15px] size-6"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              d="M17.25 8.25 21 12m0 0-3.75 3.75M21 12H3"
-            />
-          </svg>
-        </span>
-      </a>
-    </article>
-  );
-};
-
-export default WalletCard;
