--- apps/web/src/components/wallets/WalletFilters.module.scss
@@ -1,88 +1,0 @@
-.wallet-category-section {
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 0;
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 0 20px;
-  margin-bottom: 16px;
-  border-bottom: 1px solid #2b2634;
-}
-
-.wallet-category {
-  display: inline-block;
-  background-color: transparent;
-  color: #6c6a81;
-  transition: all 0.2s ease;
-  padding: 10px 16px;
-  font-size: 14px;
-  font-family: "Diatype", monotype;
-  line-height: 16px;
-  border: none;
-  border-bottom: 2px solid transparent;
-  margin-bottom: -1px;
-
-  &:hover {
-    color: #fff;
-    cursor: pointer;
-  }
-}
-
-.wallet-category--active {
-  color: #fff;
-  border-bottom-color: #9945ff;
-}
-
-.wallet-filter-section {
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 5px;
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 0 20px;
-  position: relative;
-  z-index: 10;
-}
-
-.wallet-filter {
-  display: inline-block;
-  background-color: #504d61;
-  color: #fff;
-  transition: all 0.3s ease;
-  border-radius: 30px;
-  padding: 7px 12px;
-  font-size: 12px;
-  text-transform: uppercase;
-  min-width: 50px;
-  text-align: center;
-  font-family: "Diatype", monotype;
-  line-height: 13px;
-  border: 1px solid #504d61;
-
-  &:hover {
-    background-color: #1d1a23;
-    cursor: pointer;
-  }
-}
-
-.wallet-filter--active {
-  background-color: #1d1a23;
-  color: #fff;
-}
-
-.wallet-filter + .wallet-filter-input:checked {
-  background-color: #1d1a23;
-}
-
-.wallet-filter.active {
-  background-color: #1d1a23;
-  color: #fff;
-}
-
-.wallet-filter-input {
-  appearance: none;
-}
