--- apps/web/src/components/wallets/Wallets.module.scss
@@ -1,97 +1,0 @@
-.wallets-section {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 20px;
-  margin-top: 50px;
-}
-
-.wallets-title {
-  font-size: 28px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  line-height: 33px;
-}
-
-.wallets-grid {
-  display: grid;
-  grid-template-columns: repeat(1, 1fr);
-  gap: 15px;
-  border: 1px solid #1d1a23;
-  border-radius: 20px;
-  padding: 30px;
-
-  @media all and (min-width: 768px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media all and (min-width: 1024px) {
-    grid-template-columns: repeat(4, 1fr);
-    gap: 20px;
-  }
-}
-
-.wallet {
-  display: flex;
-  flex-flow: column wrap;
-  background-color: #1d1a23;
-  border-radius: 16px;
-  padding: 30px;
-  gap: 15px;
-  cursor: pointer;
-}
-
-.wallet-card-title {
-  font-size: 28px;
-  line-height: 33px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  margin-bottom: 0;
-}
-
-.wallet-card-body {
-  font-size: 16px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  line-height: 20px;
-  color: #6c6a81;
-  margin-bottom: 0;
-}
-
-.wallet-card-body--alt {
-  font-size: 16px;
-  font-family: "Diatype", var(--font-family-sans-serif);
-  line-height: 20px;
-  margin-bottom: 0;
-  color: #fff;
-}
-
-.wallet-card-view-details {
-  font-size: 14px;
-  font-family: "Diatype", monotype;
-  line-height: 20px;
-  letter-spacing: 0.5px;
-  text-transform: uppercase;
-  text-align: left;
-  margin-bottom: 0;
-
-  span {
-    color: #9945ff;
-    font-size: 20px;
-  }
-}
-
-.wallets-modal {
-  padding: 60px 30px;
-  display: flex;
-  flex-flow: column wrap;
-  gap: 20px;
-}
-
-.wallets-no-results-container {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: flex-start;
-  align-items: center;
-  grid-column: 1/3;
-}
-
-.wallets-no-results-found {
-  margin-bottom: 0;
-}
