--- apps/web/src/components/wallets/Wallets.tsx
@@ -1,56 +1,0 @@
-import { useTranslations } from "next-intl";
-import styles from "./Wallets.module.scss";
-import Button from "../shared/Button";
-import WalletCard from "./WalletCard";
-import type { walletData } from "@workspace/ecosystem-data";
-
-type WalletEntry = (typeof walletData)[number];
-
-interface WalletsProps {
-  walletData: WalletEntry[];
-  resetWalletsAndFilters: () => void;
-}
-
-const Wallets = ({ walletData, resetWalletsAndFilters }: WalletsProps) => {
-  const t = useTranslations();
-
-  return (
-    <>
-      <section className={styles["wallets-section"]}>
-        <h2 className={styles["wallets-title"]}>{t("wallets.grid.title")}</h2>
-        <div className={styles["wallets-grid"]}>
-          {walletData.length ? (
-            walletData.map((wallet, key) => {
-              return (
-                <WalletCard
-                  index={key}
-                  name={wallet.name}
-                  walletImage={wallet.icon.src}
-                  body={wallet.body}
-                  websiteUrl={wallet.website}
-                  category={wallet.category}
-                  key={wallet.name}
-                />
-              );
-            })
-          ) : (
-            <div className={styles["wallets-no-results-container"]}>
-              <p className={styles["wallets-no-results-found"]}>
-                {t("wallets.grid.no-results")}
-              </p>
-              <Button
-                variant="outline"
-                onClick={resetWalletsAndFilters}
-                className="ml-2"
-              >
-                {t("wallets.grid.reset-filters")}
-              </Button>
-            </div>
-          )}
-        </div>
-      </section>
-    </>
-  );
-};
-
-export default Wallets;
