--- apps/web/src/data/wallets/wallet-directory.ts
@@ -1,25 +1,0 @@
-export {
-  WALLET_CATEGORIES,
-  WALLET_CATEGORY_LABELS,
-  WALLET_FEATURE_DESCRIPTIONS,
-  WALLET_FEATURE_LABELS,
-  WALLET_FEATURES,
-  WALLET_COMPANY_ALIASES,
-  WALLET_OVERRIDE_ALIASES,
-  WALLET_PLATFORM_LABELS,
-  WALLET_PLATFORMS,
-  DEFAULT_WALLET_ICON_URL,
-  curatedWalletOverrides,
-  getCuratedWalletOverride,
-  getWalletCompanyId,
-  localWalletDirectoryOverrides,
-  normalizeWalletKey,
-} from "@workspace/ecosystem-data";
-
-export type {
-  WalletCategory,
-  WalletDirectoryData,
-  WalletDirectoryEntry,
-  WalletFeature,
-  WalletPlatform,
-} from "@workspace/ecosystem-data";
