--- apps/web/src/data/wallets/wallet-filters.ts
@@ -1,82 +1,0 @@
-export const walletFiltersData = [
-  {
-    title: "Hardware",
-    filterKey: "hardware",
-    description: "",
-  },
-  {
-    title: "Custodial",
-    filterKey: "custodial",
-    description: "",
-  },
-  {
-    title: "Non-Custodial/Self-Custodial",
-    filterKey: "non_custodial",
-    description: "",
-  },
-  {
-    title: "Token Extensions",
-    filterKey: "te",
-    description: "",
-  },
-  {
-    title: "Blinks and Actions",
-    filterKey: "blinks_and_actions",
-    description: "",
-  },
-  {
-    title: "Solana Pay",
-    filterKey: "solana_pay",
-    description: "",
-  },
-  {
-    title: "Buy Crypto",
-    filterKey: "buy_crypto",
-    description: "",
-  },
-  {
-    title: "Sell Crypto",
-    filterKey: "sell_crypto",
-    description: "",
-  },
-  {
-    title: "Hold NFTs",
-    filterKey: "hold_nfts",
-    description: "",
-  },
-  {
-    title: "MPC",
-    filterKey: "mpc",
-    description: "",
-  },
-  {
-    title: "Gas Abstraction",
-    filterKey: "gas_abstraction",
-    description: "",
-  },
-  {
-    title: "Social Recovery",
-    filterKey: "social_recovery",
-    description: "",
-  },
-  {
-    title: "Staking",
-    filterKey: "staking",
-    description: "",
-  },
-  {
-    title: "Spending Limits",
-    filterKey: "spending_limits",
-    description: "",
-  },
-  {
-    title: "Open Source",
-    filterKey: "open_source",
-    description: "",
-  },
-  {
-    title: "Private Key Infrastructure",
-    filterKey: "private_key_infrastructure",
-    description: "",
-  },
-];
